Ajoute le fichier capacite_travail.py pour regrouper les variables de capacité du travail
